Added Testing for `TemtemListViewModel`.
