+++ temtemTests/ViewControllers/TemtemListViewControllerTest.swift
@@ -28,62 +28,4 @@ final class TemtemListViewControllerTest: XCTestCase {
         XCTAssertNotNil(sut.customView.tableView.delegate)
     }
     
-    func test_data_fetched() throws {
-        let sut = makeSUT()
-        let expectation = XCTestExpectation(description: "Data Fetched From API.")
-        
-        sut.viewModel.$temtems.sink { temtems in
-            XCTAssertEqual(temtems.count, 2)
-            expectation.fulfill()
-        }.store(in: &cancellables)
-        
-        wait(for: [expectation], timeout:  0.1)
-    }
-    
-    func test_search_temtem() throws {
-        let sut = makeSUT()
-        let expectation = XCTestExpectation(description: "Should have 1 searched.")
-        
-        sut.viewModel.$temtems.dropFirst().sink { temtems in
-            XCTAssertEqual(temtems.count, 1)
-            
-            expectation.fulfill()
-        }.store(in: &cancellables)
-        
-        sut.viewModel.searchTemtem(search: "golz")
-        
-        wait(for: [expectation], timeout:  0.1)
-        
-    }
-    
-    func test_empty_search_temtem() throws {
-        let sut = makeSUT()
-        let expectation = XCTestExpectation(description: "Should have Empty List.")
-        
-        sut.viewModel.$temtems.dropFirst().sink { [weak sut] temtems in
-            XCTAssertEqual(temtems.count, 0)
-            XCTAssertFalse(sut!.customView.errorLabel.isHidden)
-            
-            expectation.fulfill()
-        }.store(in: &cancellables)
-        
-        sut.viewModel.searchTemtem(search: "smazee")
-        
-        wait(for: [expectation], timeout:  0.1)
-    }
-    
-    
-    func test_empty_service() throws {
-        let sut = makeSUT(withMockService: MockTemtemServiceEmpty())
-        let expectation = XCTestExpectation(description: "Should return success but empty data.")
-        
-        sut.viewModel.$temtems.sink { temtems in
-            XCTAssertEqual(temtems.count, 0)
-            expectation.fulfill()
-        }.store(in: &cancellables)
-        
-        wait(for: [expectation], timeout:  0.1)
-        
-    }
-    
 }
