+++ temtemTests/ViewModels/TemtemListViewModelTests.swift
@@ -7,16 +7,28 @@
 
 import XCTest
 import temtem
+import Combine
 
 final class TemtemListViewModelTests: XCTestCase {
-
+    private var cancellables = Set<AnyCancellable>()
+    
     func makeSUT(file: StaticString = #filePath, line: UInt = #line) -> TemtemListViewModel {
         let sut = TemtemListViewModel(temtemService:MockTemtemService())
         testMemoryLeaks(sut: sut)
         return sut
     }
 
-    func testLaunchPerformance() throws {
-      
+    func test_data_fetched() throws {
+        let sut = makeSUT()
+        let expectation = XCTestExpectation(description: "Data Fetched From API.")
+        
+        sut.$temtems.dropFirst()
+            .sink { temtems in
+                XCTAssertEqual(temtems.count, 2)
+                expectation.fulfill()
+        }.store(in: &cancellables)
+        
+        sut.fetchTemtems()
+        wait(for: [expectation], timeout:  1)
     }
 }
